pyproject.toml
[project]
name = "pytorch-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "pillow>=12.0.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
